Added docker compose for new worker
- Added code to use the same routes as new_routes in Domain
