+++ packages/grid/.env
@@ -19,6 +19,7 @@ TRAEFIK_PUBLIC_TAG=traefik-public
 
 STACK_NAME=grid-openmined-org
 DOCKER_IMAGE_BACKEND=openmined/grid-backend
+DOCKER_IMAGE_WORKER=openmined/grid-worker
 DOCKER_IMAGE_FRONTEND=openmined/grid-frontend
 DOCKER_IMAGE_SVELTE=openmined/grid-svelte
 DOCKER_IMAGE_HEADSCALE=openmined/grid-vpn-headscale
@@ -90,8 +91,11 @@ TRACE=False
 JAEGER_HOST=localhost
 JAEGER_PORT=14268
 
-#Syft
+# Syft
 SYFT_TUTORIAL_MODE=False
+
+# Worker
+USE_BLOB_STORAGE=False
 
 #Oblivious
 INSTALL_OBLV_CLI=False
