+++ packages/grid/worker/docker-compose.build.yml
@@ -1,0 +1,9 @@
+version: "3.8"
+services:
+  worker:
+    build:
+      context: ${RELATIVE_PATH}../../
+      dockerfile: ./grid/worker/worker.dockerfile
+      target: "worker"
+      labels:
+        com.docker.compose.container-number: 1
