+++ packages/grid/worker/docker-compose.dev.yml
@@ -1,0 +1,14 @@
+version: "3.8"
+services:
+  worker:
+    volumes:
+      - ${RELATIVE_PATH}./:/app
+    #   - ${RELATIVE_PATH}./backend/worker/credentials:/credentials
+      - ${RELATIVE_PATH}../../syft/:/app/syft
+      - ${RELATIVE_PATH}../backend/grid/bootstrap.py:/app/bootstrap.py
+      - ${RELATIVE_PATH}../backend/grid/api/new/new_routes.py:/app/new_routes.py
+    #   - ${RELATIVE_PATH}./data/package-cache:/root/.cache
+    environment:
+      - DEV_MODE=True
+    ports:
+      - "$HTTP_PORT"
