+++ packages/grid/worker/docker-compose.yml
@@ -1,0 +1,34 @@
+version: "3.8"
+services:
+  worker:
+    restart: always
+    image: "${DOCKER_IMAGE_WORKER?Variable not set}:${VERSION-latest}"
+    environment:
+      - SERVICE_NAME=worker
+      - RELEASE=${RELEASE:-production}
+      - VERSION=${VERSION}
+      - VERSION_HASH=${VERSION_HASH}
+      - NODE_TYPE=${NODE_TYPE?Variable not set}
+      - DOMAIN_NAME=${DOMAIN_NAME?Variable not set}
+      - STACK_API_KEY=${STACK_API_KEY}
+      - PORT=${HTTP_PORT}
+      - IGNORE_TLS_ERRORS=${IGNORE_TLS_ERRORS?False}
+      - HTTP_PORT=${HTTP_PORT}
+      - HTTPS_PORT=${HTTPS_PORT}
+      - USE_BLOB_STORAGE=${USE_BLOB_STORAGE}
+      - CONTAINER_HOST=${CONTAINER_HOST}
+      - TRACE=${TRACE}
+      - JAEGER_HOST=${JAEGER_HOST}
+      - JAEGER_PORT=${JAEGER_PORT}
+      - ASSOCIATION_TIMEOUT=${ASSOCIATION_TIMEOUT}
+      - DEV_MODE=${DEV_MODE}
+    command: "/app/start.sh"
+    volumes:
+      - credentials-data:/credentials
+
+volumes:
+  credentials-data:
+
+networks:
+  worker-public:
+
